Move interval test to utils
